refactor(mcp): extract source test modules
- Move MCP unit tests into sibling *_test.rs files
- Keep protocol, manager, and transport source files focused on runtime code
