+++ crates/aion-mcp/src/tool_proxy.rs
@@ -188,88 +188,6 @@ pub fn register_single_server_tools(
     }
 }
 
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use aion_config::config::TransportType;
-    use serde_json::json;
-
-    fn make_proxy(deferred: bool) -> McpToolProxy {
-        // manager is only used during execute(), which we don't call in these
-        // tests, so we can construct one with no servers.
-        let manager = Arc::new(McpManager::new_for_test(vec![]));
-        McpToolProxy::new(
-            "test_tool".into(),
-            "test_tool".into(),
-            "test_server".into(),
-            "A test tool".into(),
-            json!({"type": "object"}),
-            manager,
-            deferred,
-        )
-    }
-
-    #[test]
-    fn proxy_deferred_true_returns_true() {
-        let proxy = make_proxy(true);
-        assert!(proxy.is_deferred());
-    }
-
-    #[test]
-    fn proxy_deferred_false_returns_false() {
-        let proxy = make_proxy(false);
-        assert!(!proxy.is_deferred());
-    }
-
-    fn make_server_config(deferred: Option<bool>) -> McpServerConfig {
-        McpServerConfig {
-            transport: TransportType::Stdio,
-            command: Some("echo".into()),
-            args: None,
-            env: None,
-            url: None,
-            headers: None,
-            deferred,
-            startup_timeout_ms: None,
-        }
-    }
-
-    #[test]
-    fn register_defaults_to_deferred_when_config_omits_field() {
-        let manager = Arc::new(McpManager::new_for_test(vec![]));
-        let mut registry = aion_tools::registry::ToolRegistry::new();
-        // Empty server configs — deferred field absent
-        let configs = HashMap::new();
-
-        register_mcp_tools(&mut registry, &manager, &[], &configs);
-
-        // No tools registered because manager has no tools, but the logic
-        // is tested via the deferred default path. Test with a real config below.
-        assert!(registry.tool_names().is_empty());
-    }
-
-    #[test]
-    fn server_config_deferred_none_defaults_true() {
-        let config = make_server_config(None);
-        let deferred = config.deferred.unwrap_or(true);
-        assert!(deferred, "deferred should default to true when None");
-    }
-
-    #[test]
-    fn server_config_deferred_explicit_false() {
-        let config = make_server_config(Some(false));
-        let deferred = config.deferred.unwrap_or(true);
-        assert!(!deferred, "deferred should be false when explicitly set");
-    }
-
-    #[test]
-    fn server_config_deferred_explicit_true() {
-        let config = make_server_config(Some(true));
-        let deferred = config.deferred.unwrap_or(true);
-        assert!(deferred, "deferred should be true when explicitly set");
-    }
-}
+#[path = "tool_proxy_test.rs"]
+mod tool_proxy_test;
